# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  scheduling.ErrorResponse:
    properties:
      error:
        example: failed to retrieve teachers
        type: string
    type: object
  scheduling.MessageResponse:
    properties:
      message:
        example: Availability saved successfully
        type: string
    type: object
...
  shared.Subject:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  shared.WeeklySlot:
    properties:
      day_of_week:
        description: 0=Sunday ... 6=Saturday (matches JavaScript Date.getDay())
        example: 0
        type: integer
      end:
        example: "10:00"
        type: string
      start:
        example: "09:00"
        type: string
    type: object
  teacher.CreateTeacherRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      gender:
        example: female
        type: string
      name:
        example: Alice
        type: string
    required:
    - email
    - gender
    - name
    type: object
  teacher.Teacher:
    properties:
      email:
        example: alice@example.com
        type: string
      gender:
        example: female
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Alice
        type: string
      status:
        example: active
        type: string
    type: object
  teacher.TeacherAvailability:
    properties:
      teacher:
        $ref: '#/definitions/teacher.Teacher'
      weekly:
        items:
          $ref: '#/definitions/shared.WeeklySlot'
        type: array
    type: object
  teacher.UpdateGenderRequest:
    properties:
      gender:
        example: female
        type: string
    required:
    - gender
    type: object
  teacher.UpdateStatusRequest:
    properties:
      status:
        example: deactivated
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
  description: API for teacher availability management and bookings
  title: Converge API
  version: "1.0"
paths:
  /availability:
    get:
      description: Returns all active teachers with their weekly availability slots,
        grouped by teacher
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teacher.TeacherAvailability'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      summary: Get all teacher availability
      tags:
      - availability
    post:
      consumes:
      - application/json
      description: Replaces all existing availability for a teacher with the provided
        weekly schedule (full replace, not merge). Rejects overlapping slots within
        the same day.
      parameters:
      - description: Weekly availability payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit weekly availability
      tags:
      - availability
...
  /subjects:
    get:
      description: Returns all subjects ordered by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/shared.Subject'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List subjects
      tags:
      - subjects
  /teachers:
    get:
      description: Returns teachers ordered by name. Filter by status (active|deactivated|all,
        default all). Optionally filter by subject_id, which returns active teachers
        only and ignores status.
      parameters:
      - description: Filter by subject ID (active teachers only)
        in: query
        name: subject_id
        type: integer
      - description: Filter by status
        enum:
        - active
        - deactivated
        - all
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/teacher.Teacher'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      summary: List teachers
      tags:
      - teachers
    post:
      consumes:
      - application/json
      description: Adds a new teacher with name, email, and gender
      parameters:
      - description: New teacher payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/teacher.Teacher'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a teacher
      tags:
      - teachers
  /teachers/{id}/gender:
    patch:
      consumes:
      - application/json
      description: Sets a teacher's gender to male, female, or lgbtq+
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: New gender
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateGenderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a teacher's gender
      tags:
      - teachers
  /teachers/{id}/status:
    patch:
      consumes:
      - application/json
      description: Sets a teacher's status to active or deactivated
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/teacher.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/scheduling.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle teacher active/deactivated status
      tags:
      - teachers
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT from /login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"